feat: add project badge panel to project details page (#4846)
