+++ ui/src/app/applications/components/application-summary/application-summary.scss
@@ -36,13 +36,6 @@
         right: 1em;
     }
    
-    &__badge {
-        display: inline-block;
-        width: 100%;
-        padding: .4em;
-        border: 1px solid #dbdcdc;
-        border-radius: 2px;
-    }
     .argo-field {
         line-height: 1.15;
     }
