+++ ui/src/app/applications/components/application-summary/application-summary.tsx
@@ -2,8 +2,8 @@ import {AutocompleteField, DropDownMenu, FormField, FormSelect, HelpIcon, PopupA
 import * as React from 'react';
 import {FormApi, Text} from 'react-form';
 import {Cluster, DataLoader, EditablePanel, EditablePanelItem, Expandable, MapInputField, Repo, Revision, RevisionHelpIcon} from '../../../shared/components';
-import {Spinner} from '../../../shared/components';
-import {Consumer, Context} from '../../../shared/context';
+import {BadgePanel, Spinner} from '../../../shared/components';
+import {Consumer} from '../../../shared/context';
 import * as models from '../../../shared/models';
 import {services} from '../../../shared/services';
 
@@ -378,10 +378,6 @@ export const ApplicationSummary = (props: {app: models.Application; updateApp: (
             view: null as any,
             edit: null
         });
-    const [badgeType, setBadgeType] = React.useState('URL');
-    const context = React.useContext(Context);
-    const badgeURL = `${location.protocol}//${location.host}${context.baseHref}api/badge?name=${props.app.metadata.name}&revision=true`;
-    const appURL = `${location.protocol}//${location.host}${context.baseHref}applications/${props.app.metadata.name}`;
 
     return (
         <div className='application-summary'>
@@ -498,48 +494,7 @@ export const ApplicationSummary = (props: {app: models.Application; updateApp: (
                     </div>
                 )}
             </Consumer>
-            <DataLoader load={() => services.authService.settings()}>
-                {settings =>
-                    (settings.statusBadgeEnabled && (
-                        <div className='white-box'>
-                            <div className='white-box__details'>
-                                <p>
-                                    Status Badge <img src={badgeURL} />{' '}
-                                </p>
-                                <div className='white-box__details-row'>
-                                    <DropDownMenu
-                                        anchor={() => (
-                                            <p>
-                                                {badgeType} <i className='fa fa-caret-down' />
-                                            </p>
-                                        )}
-                                        items={['URL', 'Markdown', 'Textile', 'Rdoc', 'AsciiDoc'].map(type => ({title: type, action: () => setBadgeType(type)}))}
-                                    />
-                                    <textarea
-                                        onClick={e => (e.target as HTMLInputElement).select()}
-                                        className='application-summary__badge'
-                                        readOnly={true}
-                                        value={
-                                            badgeType === 'URL'
-                                                ? badgeURL
-                                                : badgeType === 'Markdown'
-                                                ? `[![App Status](${badgeURL})](${appURL})`
-                                                : badgeType === 'Textile'
-                                                ? `!${badgeURL}!:${appURL}`
-                                                : badgeType === 'Rdoc'
-                                                ? `{<img src="${badgeURL}" alt="App Status" />}[${appURL}]`
-                                                : badgeType === 'AsciiDoc'
-                                                ? `image:${badgeURL}["App Status", link="${appURL}"]`
-                                                : ''
-                                        }
-                                    />
-                                </div>
-                            </div>
-                        </div>
-                    )) ||
-                    null
-                }
-            </DataLoader>
+            <BadgePanel app={props.app.metadata.name} />
             <EditablePanel save={props.updateApp} values={app} title='Info' items={infoItems} onModeSwitch={() => setAdjustedCount(0)} />
         </div>
     );
