+++ ui/src/app/shared/components/badge-panel/badge-panel.scss
@@ -1,0 +1,7 @@
+.badge-panel {
+    display: inline-block;
+    width: 100%;
+    padding: .4em;
+    border: 1px solid #dbdcdc;
+    border-radius: 2px;
+}
