+++ ui/src/app/shared/components/index.ts
@@ -22,3 +22,4 @@ export * from './repo';
 export * from './revision';
 export * from './timestamp';
 export * from './spinner';
+export * from './badge-panel/badge-panel';
